+++ src/server/scanImports.ts
@@ -1,4 +1,7 @@
+import { relative } from "node:path";
 import { parse } from "es-module-lexer";
+import { RDSError } from "./errors.ts";
+import type { ResolvedConfig } from "./loadConfig.ts";
 import { resolve } from "./resolve.ts";
 import { run } from "./utils.ts";
 
@@ -13,16 +16,56 @@ export type JSImport = {
   se: number;
 };
 
-export const scanImports = (url: string, code: string) => {
+export const scanImports = (
+  url: string,
+  code: string,
+  packageImports: ResolvedConfig["packageImports"],
+) => {
   const importSpecifiers = parse(code)[0];
   const localImports: JSImport[] = [];
   for (const i of importSpecifiers) {
     if (!i.n) continue; // dynamic, ex: import(variable), unhandled
     if (i.n.startsWith("https://")) continue; // external
-    const dep = !i.n.startsWith(".");
+    let r: string;
+    let dep: boolean;
+    if (i.n.startsWith("#")) {
+      const packageImport = packageImports?.[i.n];
+      if (packageImport === undefined) {
+        throw new RDSError({
+          message: `Subpath import "${i.n}" not set in package.json`,
+          file: url,
+        });
+      }
+      let subPathMapping: string;
+      if (typeof packageImport === "string") {
+        subPathMapping = packageImport;
+      } else {
+        const condition = packageImport["browser"] ?? packageImport["default"];
+        if (condition === undefined) {
+          throw new RDSError({
+            message: `Subpath import "${i.n}" condition not supported`,
+            file: url,
+          });
+        }
+        subPathMapping = condition;
+      }
+      if (subPathMapping.startsWith(".")) {
+        dep = false;
+        r = relative(process.cwd(), subPathMapping);
+      } else {
+        dep = true;
+        r = subPathMapping;
+      }
+    } else if (i.n.startsWith(".")) {
+      dep = false;
+      r = resolve(url, i.n);
+    } else {
+      dep = true;
+      r = i.n;
+    }
     localImports.push({
       n: i.n,
-      r: dep ? i.n : resolve(url, i.n),
+      r,
       dep,
       specifiers: dep
         ? run(() => {